pthread: added pthread_join(3)

+++ include/lyos/types.h
@@ -108,7 +108,7 @@ struct ps_strings {
 };
 
 struct siginfo {
-    int mask;
+    unsigned long mask;
     int signo;
     void* sig_handler;
     void* sig_return;
